requirements.txt
#LangChain ecosystem
openai == 1.35.10
langchain == 0.2.6
langchain_openai == 0.1.14
langchain_community == 0.2.6
langchain_core == 0.2.11
langchain_text_splitters == 0.2.2

#RAG loaders
pypdf == 4.2.0
jq == 1.7.0
pymssql == 2.3.0
rapidocr-onnxruntime == 1.3.23

#Vector database
faiss-cpu == 1.8.0.post1

#Deployment related
pandas
python-dotenv
psutil